Try integrate a library with a pipeline lib

+++ src/duffy.groovy
@@ -15,30 +15,33 @@ def duffy(stage, duffyKey = 'duffy-key', repoUrl = 'https://github.com/cgwalters
         }
     }
 
-    withCredentials([file(credentialsId: duffyKey, variable: 'DUFFY_KEY')]) {
-        sh '''
-            #!/bin/bash
-            set -xeuo pipefail
-    
-            cp ${DUFFY_KEY} ~/duffy.key
-            chmod 600 ~/duffy.key
+    if (duffyOps != "--no-op") {
 
-            mkdir -p ${ORIGIN_WORKSPACE}
-            # If we somehow got called without an op, do nothing.
-            if test -z "${DUFFY_OP:-}"; then
-              exit 0
-            fi
-            if test -n "${ORIGIN_WORKSPACE:-}"; then
-              pushd ${ORIGIN_WORKSPACE}
-            fi
-            if test -n "${ORIGIN_CLASS:-}"; then
-                exec ${WORKSPACE}/cciskel/cciskel-duffy ${DUFFY_OP} --prefix=ci-pipeline --class=${ORIGIN_CLASS} \
-                    --jobid=${ORIGIN_BUILD_TAG} --timeout=${DUFFY_JOB_TIMEOUT_SECS:-0} --count=${DUFFY_COUNT:-1}
-            else
-                exec ${WORKSPACE}/cciskel/cciskel-duffy ${DUFFY_OP}
-            fi
-            exit
-        '''
+        withCredentials([file(credentialsId: duffyKey, variable: 'DUFFY_KEY')]) {
+            sh '''
+                    #!/bin/bash
+                    set -xeuo pipefail
+            
+                    cp ${DUFFY_KEY} ~/duffy.key
+                    chmod 600 ~/duffy.key
+        
+                    mkdir -p ${ORIGIN_WORKSPACE}
+                    # If we somehow got called without an op, do nothing.
+                    if test -z "${DUFFY_OP:-}"; then
+                      exit 0
+                    fi
+                    if test -n "${ORIGIN_WORKSPACE:-}"; then
+                      pushd ${ORIGIN_WORKSPACE}
+                    fi
+                    if test -n "${ORIGIN_CLASS:-}"; then
+                        exec ${WORKSPACE}/cciskel/cciskel-duffy ${DUFFY_OP} --prefix=ci-pipeline --class=${ORIGIN_CLASS} \
+                            --jobid=${ORIGIN_BUILD_TAG} --timeout=${DUFFY_JOB_TIMEOUT_SECS:-0} --count=${DUFFY_COUNT:-1}
+                    else
+                        exec ${WORKSPACE}/cciskel/cciskel-duffy ${DUFFY_OP}
+                    fi
+                    exit
+                '''
+        }
     }
 }
